+++ src/main/java/com/tutti/server/core/payment/api/PaymentApi.java
@@ -1,6 +1,7 @@
 package com.tutti.server.core.payment.api;
 
 import com.tutti.server.core.payment.application.PaymentService;
+import com.tutti.server.core.payment.payload.PaymentCancelRequest;
 import com.tutti.server.core.payment.payload.PaymentConfirmRequest;
 import com.tutti.server.core.payment.payload.PaymentRequest;
 import com.tutti.server.core.payment.payload.PaymentResponse;
@@ -39,5 +40,13 @@ public class PaymentApi {
         paymentService.confirmPayment(request);
         Map<String, String> body = Map.of("message", "결제가 완료되었습니다.");
         return ResponseEntity.ok(body);
+    }
+
+    @PostMapping("/cancel")
+    public ResponseEntity<?> cancelPayment(
+            @Valid @RequestBody PaymentCancelRequest request) {
+        logger.info("결제 취소 요청: 주문 ID = {}", request.orderId());
+        paymentService.paymentCancel(request);
+        return ResponseEntity.ok(Map.of("message", "결제가 취소되었습니다."));
     }
 }
