--- src/main/java/com/tutti/server/core/payment/application/PaymentCancelSerivce.java
@@ -1,9 +1,0 @@
-package com.tutti.server.core.payment.application;
-
-import com.tutti.server.core.payment.payload.PaymentCancelRequest;
-
-public interface PaymentCancelSerivce {
-
-    void paymentCancel(PaymentCancelRequest request);
-
-}
