+++ src/main/java/com/tutti/server/core/payment/application/PaymentService.java
@@ -1,5 +1,6 @@
 package com.tutti.server.core.payment.application;
 
+import com.tutti.server.core.payment.payload.PaymentCancelRequest;
 import com.tutti.server.core.payment.payload.PaymentConfirmRequest;
 import com.tutti.server.core.payment.payload.PaymentRequest;
 import com.tutti.server.core.payment.payload.PaymentResponse;
@@ -10,5 +11,7 @@ public interface PaymentService {
     PaymentResponse requestPayment(PaymentRequest request);
 
     Map<String, Object> confirmPayment(PaymentConfirmRequest request);
+
+    void paymentCancel(PaymentCancelRequest request);
 
 }
